docs: completed data pipeline docs and refactor
