+++ data_pipeline/dags/yfinance_daily_parallel.py
@@ -11,6 +11,7 @@ from airflow.models.param import Param #type: ignore
 # Custom imports
 from utils.database import get_database_url
 from sec_data_pipeline.yfinance.yfinance_pipeline import YfinancePipeline
+from sec_data_pipeline.yfinance.yfinance_validation import YfinanceValidation
 from sec_master_db.clients.yfinance_client import YfinanceClient
 from loguru import logger
 
@@ -154,8 +155,8 @@ def yfinance_daily_parallel():
         # Deserialize DataFrame
         df = pd.read_json(StringIO(download_result['data']), orient='split')
 
-        pipeline = YfinancePipeline()
-        validation = pipeline.validate_ohlcv(df, ticker)
+        validator = YfinanceValidation()
+        validation = validator.validate_ohlcv(df, ticker)
 
         if not validation['valid']:
             failed_str = ', '.join(validation['failed_checks'])
