+++ docs/CONTRIBUTING.md
@@ -176,50 +176,3 @@ On GitHub:
 PR Title:
 - Use clear, descriptive title
 - Example: "Add portfolio analytics dashboard"
-
-PR Description Template:
-## What
-Brief description of changes
-
-## Why
-Problem this solves or feature it adds
-
-## How
-Technical approach taken
-
-## Testing
-How you tested the changes
-
-## Screenshots (if UI changes)
-[Add screenshots here]
-
-## Checklist
-- [ ] Tests pass locally
-- [ ] Documentation updated
-- [ ] No console errors
-- [ ] Followed code style
-
-### 3. PR requirements 
-
-Before your PR can be merged:
--  All commits follow conventional format (enforced by CI)
--  All GitHub Actions checks pass
--  Code review approved by @cadenlund
--  All review comments resolved
--  Branch is up to date with dev
--  PR is focused (one feature/fix only)
-
-### 4. After Approval
-
-- Only @cadenlund merges PRs
-- GitHub will prompt to delete branch - click "Delete branch"
-- Pull latest dev locally: git pull origin dev
-
-### 5. PR Labels
-
-Add relevant labels:
-- feature - New functionality
-- bug - Bug fixes
-- documentation - Docs only
-- refactor - Code improvement
-- urgent - Needs quick review
